Complete shared UI refactor
