+++ package.json
@@ -155,7 +155,8 @@
 		"trading:ui:build": "cd ui/trading && bun run build",
 		"trading:ui:serve": "bun run app:serve:trading",
 		"format:html": "bun scripts/format-html-prose.mts",
-		"knip": "bun x knip",
+		"knip": "bun x knip && bun run knip:production:trading",
+		"knip:production:trading": "bun x knip --production --workspace ui/trading --include files",
 		"knip:fix": "bun x knip --fix",
 		"ui:publish:ipfs": "docker build --target publisher -f ui/Dockerfile . -t zoltar-ui-publisher && docker run --rm --add-host=host.docker.internal:host-gateway zoltar-ui-publisher",
 		"anvil:docker": "docker run --rm --name anvil -p 8545:8545 --entrypoint anvil ghcr.io/foundry-rs/foundry:v1.5.1 --host 0.0.0.0 --port 8545 --block-base-fee-per-gas 0 --gas-price 0 --no-priority-fee --chain-id 1",
