+++ ui/coreShared/ts/tests/routing.test.ts
@@ -1,7 +1,7 @@
 /// <reference types="bun-types" />
 
 import { afterEach, beforeEach, describe, expect, test } from 'bun:test'
-import { buildRouteHref, ensureRouteHash, getCurrentRoute, getCurrentRouteHash, getRouteHash, getRouteHashSearch, getTopLevelRouteSearch, installRouting } from '../lib/routing.js'
+import { buildRouteHref, createRouting, ensureRouteHash, getCurrentRoute, getCurrentRouteHash, getRouteHash, getRouteHashSearch, getTopLevelRouteSearch, installRouting, normalizeRouteHash, parseRouteHash } from '../lib/routing.js'
 import { installDomEnvironment } from './testUtils/domEnvironment.js'
 import { installTestRouting } from './testUtils/testRouting.js'
 
@@ -58,6 +58,20 @@ describe('routing', () => {
 		expect(getCurrentRoute()).toBe('market/42')
 		window.location.hash = '#/market'
 		expect(getCurrentRoute()).toBe('not-found')
+	})
+
+	test('normalizes route hashes and resolves configured aliases through a typed router', () => {
+		const routing = createRouting({
+			defaultRoute: 'markets',
+			routes: [
+				{ aliases: ['#/developer'], hash: '#/markets', name: 'markets' },
+				{ hash: '#/portfolio', name: 'portfolio' },
+			] as const,
+		})
+		expect(parseRouteHash('#/markets?simulate=1')).toEqual({ routeHash: '#/markets', search: '?simulate=1' })
+		expect(normalizeRouteHash('markets?simulate=1')).toBe('#/markets?simulate=1')
+		expect(routing.resolve('#/developer?simulate=1')).toBe('markets')
+		expect(routing.getHash('portfolio')).toBe('#/portfolio')
 	})
 
 	test('ensureRouteHash seeds default hash when blank', () => {
