+++ ui/trading/ts/lib/routing.ts
@@ -1,4 +1,4 @@
-import { buildRouteHref, getRouteHashSearch, installRouting, type RoutingConfig } from '@zoltar/ui-core-shared/lib/routing.js'
+import { buildRouteHref, createRouting, getRouteHashSearch, installRouting, normalizeRouteHash, type RoutingConfig } from '@zoltar/ui-core-shared/lib/routing.js'
 
 export type TradingRoute = 'markets' | 'market' | 'liquidity' | 'portfolio' | 'deploy' | 'help' | `security-pool/${string}`
 
@@ -6,7 +6,7 @@ export const TRADING_ROUTING_CONFIG: RoutingConfig<TradingRoute> = {
 	defaultRoute: 'markets',
 	routes: [
 		{ hash: '#/deploy', name: 'deploy' },
-		{ hash: '#/markets', name: 'markets' },
+		{ aliases: ['#/developer'], hash: '#/markets', name: 'markets' },
 		{ hash: '#/market', name: 'market' },
 		{ hash: '#/liquidity', name: 'liquidity' },
 		{ hash: '#/portfolio', name: 'portfolio' },
@@ -20,13 +20,9 @@ export const TRADING_ROUTING_CONFIG: RoutingConfig<TradingRoute> = {
 	],
 }
 
-export function normalizeTradingRouteHash(hash: string) {
-	const queryIndex = hash.indexOf('?')
-	const routeHash = queryIndex === -1 ? hash : hash.slice(0, queryIndex)
-	const search = queryIndex === -1 ? '' : hash.slice(queryIndex)
-	const routePath = routeHash.replace(/^#\/?/, '')
-	return `${routePath === '' ? '' : `#/${routePath}`}${search}`
-}
+export const tradingRouting = createRouting(TRADING_ROUTING_CONFIG)
+
+export const normalizeTradingRouteHash = normalizeRouteHash
 
 export function getTradingRouteHref(routeHash: string) {
 	return buildRouteHref(routeHash, getRouteHashSearch())
