+++ ui/trading/ts/lib/walletSummaryState.ts
@@ -7,3 +7,17 @@ export type WalletSummaryState = Readonly<{
 	errorLabel: string | undefined
 	universeId: string | undefined
 }>
+
+export function walletSummaryForUniverse(summary: WalletSummaryState, selectedUniverseId: string | undefined): WalletSummaryState {
+	if (summary.universeId === selectedUniverseId) return summary
+	return { account: summary.account, ethAttoEth: undefined, repAttoRep: undefined, status: summary.account === undefined ? 'disconnected' : 'loading', error: undefined, errorLabel: undefined, universeId: selectedUniverseId }
+}
+
+export function routeOwnsLiveWallet(route: string) {
+	return route !== 'deploy' && route !== 'help' && route !== 'not-found'
+}
+
+export function walletSummaryAfterRouteChange(summary: WalletSummaryState, previousRoute: string, nextRoute: string, selectedUniverseId: string | undefined): WalletSummaryState {
+	if (routeOwnsLiveWallet(previousRoute) === routeOwnsLiveWallet(nextRoute)) return summary
+	return { account: undefined, ethAttoEth: undefined, repAttoRep: undefined, status: 'disconnected', error: undefined, errorLabel: undefined, universeId: selectedUniverseId }
+}
