+++ ui/trading/ts/tests/app/universe-selector.test.tsx
@@ -1,7 +1,11 @@
 import { afterEach, beforeEach, describe, expect, test } from 'bun:test'
 import { act } from 'preact/test-utils'
 import { installDomEnvironment } from '@zoltar/ui-core-shared/tests/testUtils/domEnvironment.js'
-import { App, buildLiveUniverseOptions, compactUniqueUniverseIds, currentRoute, routeOwnsLiveWallet, tradingDocumentTitle, UniverseSelector, WalletSummary, walletSummaryAfterRouteChange, walletSummaryForUniverse } from '../../app/App.js'
+import { App, currentRoute, tradingDocumentTitle } from '../../app/App.js'
+import { UniverseSelector } from '../../components/UniverseSelector.js'
+import { WalletSummary } from '../../components/WalletSummary.js'
+import { buildLiveUniverseOptions, compactUniqueUniverseIds } from '../../lib/universeOptions.js'
+import { routeOwnsLiveWallet, walletSummaryAfterRouteChange, walletSummaryForUniverse } from '../../lib/walletSummaryState.js'
 import { filterMarketsByUniverse, observeKnownReceipt, walletSummaryAvailability, walletSummaryDiscoveryRetryStart, walletSummaryRefreshState } from '../../features/LiveTrading.js'
 import type { DeploymentConfiguration } from '../../protocol/config.js'
 import type { LiveMarket } from '../../protocol/live.js'
