+++ ui/trading/ts/tests/features/live-workflow-safety.test.tsx
@@ -4,7 +4,8 @@ import { act } from 'preact/test-utils'
 import type { Address, Hash } from '@zoltar/shared/ethereum'
 import { installDomEnvironment } from '@zoltar/ui-core-shared/tests/testUtils/domEnvironment.js'
 import type { DeploymentConfiguration } from '../../protocol/config.js'
-import { LiveTrading as ProductionLiveTrading, liveLiquidityServices, liveSettlementServices, type WalletSummaryState } from '../../features/LiveTrading.js'
+import { LiveTrading as ProductionLiveTrading, liveLiquidityServices, liveSettlementServices } from '../../features/LiveTrading.js'
+import type { WalletSummaryState } from '../../lib/walletSummaryState.js'
 import { liveTradingControllerServices } from '../../features/liveTradingController.js'
 import * as actualLive from '../../protocol/live.js'
 import type { LiveMarket } from '../../protocol/live.js'
