+++ ui/trading/ts/tests/features/portfolio-scope.test.tsx
@@ -2,8 +2,6 @@ import { afterEach, beforeEach, describe, expect, test } from 'bun:test'
 import type { Address } from '@zoltar/shared/ethereum'
 import { installDomEnvironment } from '@zoltar/ui-core-shared/tests/testUtils/domEnvironment.js'
 import { LivePortfolio, LiveSecurityPoolDetails, PairInitializationAction, SecurityPoolRouteEmptyState } from '../../features/LiveTrading.js'
-import { Portfolio } from '../../features/Routes.js'
-import { demoMarket } from '../../demo/markets.js'
 import type { LiveMarket } from '../../protocol/live.js'
 import { renderIntoDocument } from '@zoltar/ui-core-shared/tests/testUtils/renderIntoDocument.js'
 
@@ -98,16 +96,6 @@ describe('live portfolio scope', () => {
 		expect(rendered.container.textContent).not.toContain('These balances and LP claims belong only')
 		expect(rendered.container.textContent).not.toContain('live RPC')
 		expect(rendered.container.textContent).not.toContain('Balances are grouped by SecurityPool')
-	})
-
-	test('shows demo balances only for the globally selected universe', async () => {
-		const rendered = await renderIntoDocument(<Portfolio market={demoMarket('baseline')} />)
-		cleanupRendered = rendered.cleanup
-		expect(rendered.container.querySelectorAll('[data-portfolio-pool]')).toHaveLength(1)
-		expect(rendered.container.textContent).not.toContain('Genesis universe')
-		expect(rendered.container.textContent).not.toContain('Child universe · YES branch')
-		expect(rendered.container.textContent).not.toContain('Total YES')
-		expect(rendered.container.textContent).not.toContain('These balances and LP claims belong only')
 	})
 
 	test('keeps live pool identifiers and operational details in the security pool view', async () => {
